ds private distributing specific
