+++ src/ds/ds-color-common.h
@@ -5,7 +5,7 @@
 
 #include "device.h"
 #include "core/video.h"
-#include "ds/ds-device-common.h"
+#include "ds-device-common.h"
 
 namespace librealsense
 {
@@ -16,7 +16,7 @@ namespace librealsense
             synthetic_sensor& color_ep,
             firmware_version fw_version,
             std::shared_ptr<hw_monitor> hw_monitor,
-            device* owner, ds_device_type device_type);
+            device* owner, ds::ds_device_type device_type);
         void register_color_options();
         void register_standard_options();
         void register_metadata();
@@ -27,7 +27,7 @@ namespace librealsense
         firmware_version _fw_version;
         std::shared_ptr<hw_monitor> _hw_monitor;
         device* _owner;
-        ds_device_type _device_type;
+        ds::ds_device_type _device_type;
     };
