+++ src/ds/ds-motion-common.h
@@ -5,6 +5,8 @@
 
 #include "ds-device-common.h"
 #include "core/video.h"
+#include "ds5/ds5-private.h"
+#include "ds6/ds6-private.h"
 
 namespace librealsense
 {
@@ -415,7 +417,7 @@ namespace librealsense
     {
     public:
         explicit ds_fisheye_sensor(std::shared_ptr<sensor_base> sensor,
-            device* owner, ds_device_type device_type);
+            device* owner, ds::ds_device_type device_type);
 
         rs2_intrinsics get_intrinsics(const stream_profile& profile) const override;
         stream_profiles init_stream_profiles() override;
@@ -426,7 +428,7 @@ namespace librealsense
         std::shared_ptr<stream_interface> get_fisheye_stream() const;
 
         device* _owner;
-        ds_device_type _device_type;
+        ds::ds_device_type _device_type;
     };
 
     class ds_hid_sensor : public synthetic_sensor,
@@ -435,7 +437,7 @@ namespace librealsense
     public:
         explicit ds_hid_sensor(std::string name,
             std::shared_ptr<sensor_base> sensor,
-            device* owner, ds_device_type device_type);
+            device* owner, ds::ds_device_type device_type);
 
         rs2_motion_device_intrinsic get_motion_intrinsics(rs2_stream stream) const;
 
@@ -446,7 +448,7 @@ namespace librealsense
         std::shared_ptr<stream_interface> get_gyro_stream() const;
 
         const device* _owner;
-        ds_device_type _device_type;
+        ds::ds_device_type _device_type;
     };
 
     class global_time_option;
@@ -456,7 +458,7 @@ namespace librealsense
     {
     public:
         ds_motion_common(device* owner,
-            ds_device_type device_type,
+            ds::ds_device_type device_type,
             firmware_version fw_version,
             const ds::d400_caps& device_capabilities,
             std::shared_ptr<hw_monitor> hwm);
@@ -499,12 +501,14 @@ namespace librealsense
         void set_roi_method();
         const stream_interface& get_depth_stream() const;
         void register_streams_to_extrinsic_groups();
+        std::vector<platform::uvc_device_info> filter_device_by_capability(const std::vector<platform::uvc_device_info>& devices,
+            ds::d400_caps caps);
 
         friend class ds_hid_sensor;
         friend class ds_fisheye_sensor;
 
         device* _owner;
-        ds_device_type _ds_device_type;
+        ds::ds_device_type _ds_device_type;
         firmware_version _fw_version;
         ds::d400_caps _device_capabilities;
         std::shared_ptr<hw_monitor> _hw_monitor;
