+++ src/ds/ds5/ds5-color.cpp
@@ -105,7 +105,7 @@ namespace librealsense
         auto& raw_color_ep = get_raw_color_sensor();
     
         _ds_color_common = std::make_shared<ds_color_common>(raw_color_ep, color_ep, 
-            _fw_version, _hw_monitor, this, ds_device_type::ds5);
+            _fw_version, _hw_monitor, this, ds::ds_device_type::ds5);
 
         register_options();
         register_metadata();
