+++ src/ds/ds6/ds6-private.cpp
@@ -9,6 +9,47 @@ namespace librealsense
 {
     namespace ds
     {
+        bool ds6_try_fetch_usb_device(std::vector<platform::usb_device_info>& devices,
+            const platform::uvc_device_info& info, platform::usb_device_info& result)
+        {
+            for (auto it = devices.begin(); it != devices.end(); ++it)
+            {
+                if (it->unique_id == info.unique_id)
+                {
+                    bool found = false;
+                    result = *it;
+                    switch (info.pid)
+                    {
+                    case RS_D585_PID:
+                    case RS_S585_PID:
+                        found = (result.mi == 6);
+                        break;
+                    default:
+                        throw not_implemented_exception(to_string() << "USB device "
+                            << std::hex << info.pid << ":" << info.vid << std::dec << " is not supported.");
+                        break;
+                    }
 
+                    if (found)
+                    {
+                        devices.erase(it);
+                        return true;
+                    }
+                }
+            }
+            return false;
+        }
+
+        std::vector<platform::uvc_device_info> filter_ds6_device_by_capability(const std::vector<platform::uvc_device_info>& devices,
+            d400_caps caps)
+        {
+            std::vector<platform::uvc_device_info> results;
+
+            throw invalid_value_exception(to_string()
+                << "Capability filters are not implemented for val "
+                << std::hex << caps << std::dec);
+
+            return results;
+        }
     } // librealsense::ds
 } // namespace librealsense
