+++ src/media/ros/ros_reader.h
@@ -85,7 +85,7 @@ namespace librealsense
         bool is_color_sensor(std::string sensor_name);
         bool is_motion_module_sensor(std::string sensor_name);
         bool is_fisheye_module_sensor(std::string sensor_name);
-        bool is_ds5_PID(int pid);
+        bool is_ds_PID(int pid);
         bool is_sr300_PID(int pid);
         bool is_l500_PID(int pid);
         std::shared_ptr<recommended_proccesing_blocks_snapshot> read_proccesing_blocks_for_version_under_4(std::string pid, std::string sensor_name, std::shared_ptr<options_interface> options);
